+++ app/controllers/client_session_controller.rb
@@ -6,7 +6,6 @@ class ClientSessionController < ApplicationController
 
   # Show the client login page
   def new
-    redirect_to client_project_root_path(current_client) if current_client
   end
 
   # Attempt to login the client
@@ -14,10 +13,16 @@ class ClientSessionController < ApplicationController
     project = Project.current.where.not(username: [nil,""]).find_by_username(params[:client][:username])
 
     if project && project.authenticate(params[:client][:password])
-      session[:project_id] = project.id
+      session[:project_ids] ||= []
+      session[:project_ids] << project.id
       flash[:alert] = nil
       flash[:notice] = nil
-      redirect_to client_project_menu_path(project)
+
+      if project.to_param == session[:last_project_param]
+        redirect_to (session[:previous_client_url] || client_project_menu_path(project))
+      else
+        redirect_to client_project_menu_path(project)
+      end
     else
       flash[:alert] = 'Invalid username or password.'
       render :new
@@ -26,7 +31,8 @@ class ClientSessionController < ApplicationController
 
   # Logout the client
   def destroy
-    session[:project_id] = nil
+    session[:previous_client_url] = session[:last_project_param] = nil
+    session[:project_ids] = nil
     redirect_to client_login_path
   end